Run dev app server directly from runfiles tree
Previously, these rules were zipping up a .war and then, when the
executable ran, it would make a temporary directory and unzip the war
there. It is faster to develop just using the runfiles tree.

This commit also makes several other changes (that I can split out if
you prefer):

* Doesn’t include every jar in the app engine sdk in the build (only
the ones that, AFAICT, you actually need).
* Fails fast if the app engine sdk glob didn’t return anything (which
happens to me a lot when I use a local_repo for
com_google_appengine_java because I always forget it expects an
APPENGINE_DIR subdirectory).
* Uses .format for python string replacement